backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
pydantic[email]==2.10.3
pydantic-settings==2.7.0
aiosqlite==0.20.0
google-generativeai==0.8.3
python-dateutil==2.9.0
